requirements.txt
certifi==2020.11.8
chardet==3.0.4
colored==1.4.2
idna==2.10
python-dotenv==0.15.0
requests==2.25.0
urllib3==1.26.2
